+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v ?.??.?.? - ?? ??? ????
 - MD5Materials can be read from a PATH location
+- Expanded UBO support to more cleanly set uniforms within buffer block
 
 ## v 5.25.0.0 - 09 Dec 2025
 - MD5Models support normal mapping